// gitlabdockercompose-main/app/types/errors/codes.ts
export const enum ErrorCodes {
  E401 = 401,
  E1001 = 1001,
  E1002 = 1002,
  E1003 = 1003,
  E1004 = 1004,
  E1005 = 1005,
  E1006 = 1006,
  E1007 = 1007,
  E1008 = 1008,
  E1009 = 1009,
  E1010 = 1010,
  E1011 = 1011,
  E1012 = 1012,
  E1013 = 1013,
  E1014 = 1014,
  E2001 = 2001,
  E6301 = 6301,
}